+++ src/spawn/ast_impl/top_level_impl.rs
@@ -1,8 +1,9 @@
 use env::{ArgumentsEnvironment, AsyncIoEnvironment, ExecutableEnvironment,
           ExportedVariableEnvironment, FileDescEnvironment, FileDescOpener, FunctionEnvironment,
           IsInteractiveEnvironment, LastStatusEnvironment, ReportFailureEnvironment,
-          SetArgumentsEnvironment, StringWrapper, SubEnvironment, UnsetVariableEnvironment,
-          WorkingDirectoryEnvironment};
+          RedirectRestorer, SetArgumentsEnvironment, StringWrapper, SubEnvironment,
+          UnsetVariableEnvironment, VarRestorer, WorkingDirectoryEnvironment};
+use env::builtin::{BuiltinEnvironment, BuiltinUtility};
 use error::RuntimeError;
 use eval::{Fields, WordEval, WordEvalConfig};
 use failure::Fail;
@@ -13,14 +14,18 @@ use spawn::{BoxSpawnEnvFuture, BoxStatusFuture, Spawn, SpawnBoxed};
 use std::fmt::Display;
 use std::rc::Rc;
 use std::sync::Arc;
+use std::vec::IntoIter;
 use conch_parser::ast::{AtomicTopLevelCommand, AtomicTopLevelWord, TopLevelCommand, TopLevelWord};
 
 macro_rules! impl_top_level_cmd {
     ($type: ident, $Rc:ident, $($extra_bounds:tt)*) => {
-        impl<T, E: ?Sized> Spawn<E> for $type<T>
+        impl<T, B, PB, E: ?Sized> Spawn<E> for $type<T>
             where T: 'static + StringWrapper + Display $($extra_bounds)*,
+                  B: BuiltinUtility<IntoIter<T>, RedirectRestorer<E>, VarRestorer<E>, PreparedBuiltin = PB>,
+                  PB: Spawn<E>,
                   E: 'static + AsyncIoEnvironment
                     + ArgumentsEnvironment<Arg = T>
+                    + BuiltinEnvironment<BuiltinName = <E as FunctionEnvironment>::FnName, Builtin = B>
                     + ExecutableEnvironment
                     + ExportedVariableEnvironment<VarName = T, Var = T>
                     + FileDescEnvironment
@@ -39,7 +44,8 @@ macro_rules! impl_top_level_cmd {
                   E::Fn: Clone
                     + From<$Rc<'static + SpawnBoxed<E, Error = RuntimeError> $($extra_bounds)*>>
                     + Spawn<E, Error = RuntimeError>,
-                  <E::Fn as Spawn<E>>::Error: From<<E::ExecFuture as Future>::Error>,
+                  <E::Fn as Spawn<E>>::Error: From<<E::ExecFuture as Future>::Error>
+                      + From<PB::Error>,
                   <E::ExecFuture as Future>::Error: Fail,
                   E::IoHandle: From<E::FileHandle> + From<E::OpenedFileHandle>,
         {
@@ -52,10 +58,13 @@ macro_rules! impl_top_level_cmd {
             }
         }
 
-        impl<'a, T: 'a, E: ?Sized> Spawn<E> for &'a $type<T>
+        impl<'a, T: 'a, B, PB, E: ?Sized> Spawn<E> for &'a $type<T>
             where T: 'static + StringWrapper + Display $($extra_bounds)*,
+                  B: BuiltinUtility<IntoIter<T>, RedirectRestorer<E>, VarRestorer<E>, PreparedBuiltin = PB>,
+                  PB: Spawn<E>,
                   E: 'static + AsyncIoEnvironment
                     + ArgumentsEnvironment<Arg = T>
+                    + BuiltinEnvironment<BuiltinName = <E as FunctionEnvironment>::FnName, Builtin = B>
                     + ExecutableEnvironment
                     + ExportedVariableEnvironment<VarName = T, Var = T>
                     + FileDescEnvironment
@@ -74,7 +83,8 @@ macro_rules! impl_top_level_cmd {
                   E::Fn: Clone
                     + From<$Rc<'static + SpawnBoxed<E, Error = RuntimeError> $($extra_bounds)*>>
                     + Spawn<E, Error = RuntimeError>,
-                  <E::Fn as Spawn<E>>::Error: From<<E::ExecFuture as Future>::Error>,
+                  <E::Fn as Spawn<E>>::Error: From<<E::ExecFuture as Future>::Error>
+                      + From<PB::Error>,
                   <E::ExecFuture as Future>::Error: Fail,
                   E::IoHandle: From<E::FileHandle> + From<E::OpenedFileHandle>,
         {
@@ -91,10 +101,13 @@ macro_rules! impl_top_level_cmd {
 
 macro_rules! impl_top_level_word {
     ($type:ident, $Rc:ident, $($extra_bounds:tt)*) => {
-        impl<T, E: ?Sized> WordEval<E> for $type<T>
+        impl<T, B, PB, E: ?Sized> WordEval<E> for $type<T>
             where T: 'static + StringWrapper + Display $($extra_bounds)*,
+                  B: BuiltinUtility<IntoIter<T>, RedirectRestorer<E>, VarRestorer<E>, PreparedBuiltin = PB>,
+                  PB: Spawn<E>,
                   E: 'static + AsyncIoEnvironment
                     + ArgumentsEnvironment<Arg = T>
+                    + BuiltinEnvironment<BuiltinName = <E as FunctionEnvironment>::FnName, Builtin = B>
                     + ExecutableEnvironment
                     + ExportedVariableEnvironment<VarName = T, Var = T>
                     + FileDescEnvironment
@@ -113,7 +126,8 @@ macro_rules! impl_top_level_word {
                   E::Fn: Clone
                     + From<$Rc<'static + SpawnBoxed<E, Error = RuntimeError> $($extra_bounds)*>>
                     + Spawn<E, Error = RuntimeError>,
-                  <E::Fn as Spawn<E>>::Error: From<<E::ExecFuture as Future>::Error>,
+                  <E::Fn as Spawn<E>>::Error: From<<E::ExecFuture as Future>::Error>
+                      + From<PB::Error>,
                   <E::ExecFuture as Future>::Error: Fail,
                   E::IoHandle: From<E::FileHandle> + From<E::OpenedFileHandle>,
         {
@@ -126,10 +140,13 @@ macro_rules! impl_top_level_word {
             }
         }
 
-        impl<'a, T, E: ?Sized> WordEval<E> for &'a $type<T>
+        impl<'a, T, B, PB, E: ?Sized> WordEval<E> for &'a $type<T>
             where T: 'static + StringWrapper + Display $($extra_bounds)*,
+                  B: BuiltinUtility<IntoIter<T>, RedirectRestorer<E>, VarRestorer<E>, PreparedBuiltin = PB>,
+                  PB: Spawn<E>,
                   E: 'static + AsyncIoEnvironment
                     + ArgumentsEnvironment<Arg = T>
+                    + BuiltinEnvironment<BuiltinName = <E as FunctionEnvironment>::FnName, Builtin = B>
                     + ExecutableEnvironment
                     + ExportedVariableEnvironment<VarName = T, Var = T>
                     + FileDescEnvironment
@@ -148,7 +165,8 @@ macro_rules! impl_top_level_word {
                   E::Fn: Clone
                     + From<$Rc<'static + SpawnBoxed<E, Error = RuntimeError> $($extra_bounds)*>>
                     + Spawn<E, Error = RuntimeError>,
-                  <E::Fn as Spawn<E>>::Error: From<<E::ExecFuture as Future>::Error>,
+                  <E::Fn as Spawn<E>>::Error: From<<E::ExecFuture as Future>::Error>
+                      + From<PB::Error>,
                   <E::ExecFuture as Future>::Error: Fail,
                   E::IoHandle: From<E::FileHandle> + From<E::OpenedFileHandle>,
         {